# services/indexer/openapi.yaml
openapi: "3.1.0"
info:
  title: Linkora Indexer API
  version: "0.1.0"
  description: REST API for querying indexed on-chain data from the Linkora protocol.
servers:
  - url: http://localhost:3001/api
    description: Local development server
paths:
...
  /posts:
    get:
      summary: List posts
      operationId: listPosts
      tags: [Posts]
      parameters:
        - name: author
          in: query
          required: false
          schema:
            type: string
          description: Filter posts by author Stellar address.
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: Paginated list of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
        "400":
          $ref: "#/components/responses/ValidationError"
...
components:
  parameters:
...
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Cursor for pagination.
  schemas:
...
    Post:
      type: object
      required: [id, author, deleted, tip_total, like_count, created_ledger]
      properties:
        id:
          type: string
          description: Numeric post ID serialised as a string.
        author:
          type: string
        deleted:
          type: boolean
        tip_total:
          type: string
          description: Cumulative tip amount serialised as a string.
        like_count:
          type: string
          description: Total likes serialised as a string.
        created_ledger:
          type: integer
        deleted_ledger:
          type: integer
          nullable: true
    PostList:
      type: object
      required: [posts, total, limit, cursor, has_more]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        limit:
          type: integer
        cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code from the taxonomy (VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN, CONFLICT, RATE_LIMITED, INTERNAL_ERROR, SERVICE_UNAVAILABLE).
        message:
          type: string
          description: Human-readable error message.
        details:
          type: array
          description: Optional additional context (e.g. validation errors).
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
        requestId:
          type: string
          description: Request correlation ID for tracing.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
  responses:
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request data
              details:
                - path: limit
                  message: Expected integer, received string
              requestId: abc123
...